38 Move helper function into seperate file.
